% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractNode.R
\docType{data}
\name{AbstractNode}
\alias{AbstractNode}
\title{AbstractNode.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractNode
}
\description{
AbstractNode is an abstract class.
It defines the interfaces required for subclasses,
that must be compatible as a node in a CompositeTree.
Its basic definition is that it comprises:
- a number of input bits,
- a number of output bits,
- a method to run its algorithm for a given input and return its output,
- utility methods of general usage.
}
\keyword{datasets}
