% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_algo_base.R
\docType{data}
\name{algo_base}
\alias{algo_base}
\title{algo_base (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
algo_base
}
\description{
algo_base is the base abstract class that implements the concept of algorithm in \eqn{\mathbb{A}^{\mathbb{B}}}.
Instances of classes that implement algo_base may act as a component in an algorithm composite.
In addition to the formal components expected from the definition of an algorithm in \eqn{\mathbb{A}^{\mathbb{B}}}, the class provides complementary services such as unique IDs, labels, etc.
}
\seealso{
\code{\link{algorithms_in_ab}}

\code{\link{notation}}
}
\keyword{datasets}
