% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_algo_composite.R
\docType{data}
\name{algo_composite}
\alias{algo_composite}
\title{algo_composite (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
algo_composite
}
\description{
A algo_composite is an algorithm that is composed of sub-algorithms.
A algo_composite is itself of type algo_base, enabling complex and deep trees.
}
\keyword{datasets}
