% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AlgoComposite.R
\docType{data}
\name{AlgoComposite}
\alias{AlgoComposite}
\title{AlgoComposite}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AlgoComposite
}
\description{
A AlgoComposite is an algorithm that is composed of sub-algorithms.
A AlgoComposite is itself of type AlgoNode, enabling complex and deep trees.
}
\keyword{datasets}
