% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgoNAND.R
\docType{data}
\name{AlgoNAND}
\alias{AlgoNAND}
\title{AlgoNAND}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AlgoNAND
}
\description{
The atomic NAND algorithm node.
}
\keyword{datasets}
