% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AlgoNode.R
\docType{data}
\name{AlgoNode}
\alias{AlgoNode}
\title{AlgoNode}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AlgoNode
}
\description{
AlgoNode is an abstract class that should be inherited from.
It defines the interfaces required for objects that must be compatible as a node in a CompositeTree.
Its basic definition is that it comprises:
- identification, labeling and other decorative information,
- a number of input bits,
- a number of output bits,
- a method to run its algorithm for a given input and return its output,
- utility methods of general usage.
}
\keyword{datasets}
