% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_algorithms_in_ab.R
\name{algorithms_in_ab}
\alias{algorithms_in_ab}
\title{Algorithms in \eqn{\mathbb{A}^{\mathbb{B}}}}
\description{
An algorithm \eqn{a} in \eqn{\mathbb{A}^{\mathbb{B}}}, is defined as a structure composed of:
\itemize{
\item{ An ordered set of \eqn{i} input bits \eqn{i_{1}, i_{2}, ..., i_{i}} }
\item{ An ordered set of \eqn{j} output bits \eqn{o_{1}, o_{2}, ..., o_{j}} }
\item{ A mechanism that maps all input bit values to output bit values }
}
}
\seealso{
\code{\link{algo_base}}

\code{\link{notation}}
}
