% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AlgoTT.R
\docType{data}
\name{AlgoTT}
\alias{AlgoTT}
\title{AlgoTT}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AlgoTT
}
\description{
To store and manage truthtables with an input of arbitrary size N and an output of arbitrary size M, we will use a logical matrix with size 2 ^ N rows and M columns.
It is not necessary to store the input of the truthtable in the truthtable data structure because we will use the row index position of the output to infer the input. For instance:
Row index position = 1 --> Input = binum of value 0.
Row index position = 2 --> Input = binum of value 1.
Row index position = 3 --> Input = binum of value 2.
...
Row index position = 2 ^ N --> Input = binum of value 2 ^ N - 1.
}
\examples{
t1 <- AlgoTT$new(
  input_dimension = 2,
  output_dimension = 2);
t1$set_output(input = "01", output = "11");
t1$do_execute("01");

}
\keyword{datasets}
