% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_atomize.R
\name{atomize}
\alias{atomize}
\title{atomize}
\usage{
atomize(algo, ...)
}
\arguments{
\item{algo}{The original algorithm (R6 Class algo_base)}

\item{...}{For future usage}
}
\value{
A composite algorithm (R6 Class algo_composite)
}
\description{
This transformation takes a composite or truth table in input, \cr
and returns a composite algorithm with equivalent logic,
that is only composed of atomic elements (NANDs and bit constants).
}
\examples{
...

}
