% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bdom.R
\docType{data}
\name{bdom}
\alias{bdom}
\title{bdom (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bdom
}
\arguments{
\item{dimension}{The dimension of the modular binary domain. (integer)}

\item{...}{For future usage.}
}
\value{
An object instance of R6 class \code{bdom}.
}
\description{
\code{bdom} is an R6 implementation of the modular binary domain concept.
}
\section{References}{

\itemize{
\item{Modular binary domain (\link{def_modular_binary_domain})}
}
}

\examples{
bd1 <- bdom$new(1);
print(bd1);
bd2 <- bdom$new(2);
print(bd2);
bd3 <- bdom$new(3);
print(bd3);

}
\keyword{datasets}
