% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bnum.R
\docType{data}
\name{bnum}
\alias{bnum}
\title{bnum (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bnum
}
\arguments{
\item{input}{(conditional) An initialization value (character | logical vector)}

\item{dim}{(conditional) The desired dimension of the binary number (integer)}

\item{...}{For future usage.}
}
\value{
An object instance of class bnum.
}
\description{
The class implementation of the binary number concept. \cr
}
\examples{
# Construction from character string
n1 <- bnum$new("0101111000");
n1$plot();

# Construction from logical vector
n2 <- bnum$new(c(TRUE,FALSE,FALSE,TRUE));
n2$plot();

# Construction with dimension, defaulting to 0
n3 <- bnum$new(dim = 5);
n3$plot();

}
\keyword{datasets}
