% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_equal_bnum.R
\name{equal_bnum}
\alias{equal_bnum}
\title{equal_bnum}
\usage{
equal_bnum(n1, n2)
}
\arguments{
\item{n1}{A modular binary number (logical vector, character vector, bnum)}

\item{n2}{A modular binary number (logical vector, character vector, bnum)}
}
\value{
Returns TRUE if n1 and n2 are equal, FALSE otherwise (logical vector of length 1)
}
\description{
Tells us if two bnum objects are equal.
}
\examples{
equal_bnum(bnum$new("1101"), bnum$new("110"));
equal_bnum(bnum$new("1101"), bnum$new("1100"));
equal_bnum(bnum$new("1101"), bnum$new("1101"));
equal_bnum(bnum$new("000"), bnum$new("0000"));
equal_bnum(bnum$new("000"), bnum$new("000"));

}
