% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bset.R
\docType{data}
\name{bset}
\alias{bset}
\title{bset (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bset
}
\arguments{
\item{dimension}{The dimension of the modular binary domain. (integer)}

\item{...}{For future usage.}
}
\value{
An object instance of R6 class \code{bset}.
}
\description{
\code{bset} is the R6 class implementation of the \link[modular_binary_domain]{modular binary domain} concept.
}
\section{References}{

\itemize{
\item{\link[modular_binary_domain]{modular binary domain}}
}
}

\examples{
bd1 <- bset$new(1);
print(bd1);
bd2 <- bset$new(2);
print(bd2);
bd3 <- bset$new(3);
print(bd3);

}
\keyword{datasets}
