% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_commutate.R
\name{commutate}
\alias{commutate}
\title{commutate}
\usage{
commutate(algo_0, algo_1, ...)
}
\arguments{
\item{algo_0}{A component algorithm to be commutated to when e == 0 (algo_base)}

\item{algo_1}{A component algorithm to be commutated to when e == 1 (algo_base)}

\item{...}{For future usage}
}
\value{
A component algorithm whose inner logic commutates algo_0 and algo_1 based on e
}
\description{
...
}
\examples{
# Switch two simple algos:
require(haricot);
algo_0 <- algo_01$new();
algo_0$plot();
algo_1 <- algo_10$new();
algo_1$plot();
commutated <- commutate(algo_0, algo_1);
commutated$exec("00");
commutated$exec("10");
commutated$exec("01");
commutated$exec("11");
commutated$plot();

}
