% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositeAlgo.R
\docType{data}
\name{CompositeAlgo}
\alias{CompositeAlgo}
\title{CompositeAlgo}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompositeAlgo
}
\description{
A CompositeAlgo is an algorithm that is composed of sub-algorithms.
A CompositeAlgo is itself an CompositeAlgoInnerNode, enabling complex and deep trees.
}
\keyword{datasets}
