% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositeDiGraph.R
\docType{data}
\name{CompositeDiGraph}
\alias{CompositeDiGraph}
\title{CompositeDiGraph}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompositeDiGraph
}
\description{
A CompositeDiGraph is an algorithm that is composed of AbstractNodes.
A CompositeDiGraph is itself an AbstractNode, enabling complex and deep trees.
}
\keyword{datasets}
