% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositeTree.R
\docType{data}
\name{CompositeTree}
\alias{CompositeTree}
\title{CompositeTree}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompositeTree
}
\description{
A CompositeTree is an algorithm that is composed of AbstractNodes.
A CompositeTree should itself be a AbstractNode, enabling complex and deep trees.
}
\keyword{datasets}
