% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_concat_truthtable.R
\name{concatenate_truthtable}
\alias{concatenate_truthtable}
\title{Concatenates two truthtables of equal input dimension}
\usage{
concatenate_truthtable(truthtable_1, truthtable_2)
}
\arguments{
\item{truthtable_1}{The first truthtable of size x (TruthTable)}

\item{truthtable_2}{The second truthtable of size x (TruthTable)}
}
\value{
A truthtable of size x (TruthTable)
}
\description{
By concatenation in this context,
we mean merging together two algorithms/functions,
in such a way that the output of the merge algorithm/function
is equal to the output of the first truthtable
followed by the output of the second truthtable.
The two truthtables must have an identical input dimension.
}
