% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_constant.R
\name{constant_algorithm}
\alias{constant_algorithm}
\title{Constant algorithm}
\description{
By definition, a constant algorithm is an algorithm with an input dimension of 0. \cr
The rationale behind this definition is that a deterministic algorithm that receives no input must mechanically output the same value independently of the (non-existent) input.
}
