% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_algo_base_to_igraph.R
\name{convert_algo_base_to_igraph}
\alias{convert_algo_base_to_igraph}
\title{Convert an object of class algo_base to an object of class igraph.}
\usage{
convert_algo_base_to_igraph(node, ...)
}
\arguments{
\item{node}{The node that we want to graph (algo_base)}

\item{...}{For future usage.}
}
\value{
A directed network graph representation of the node (igraph).
}
\description{
The igraph network will be composed of:
* One node per input bit,
* One node representing the algorithm,
* One node per output bit,
* Directed edges from the input bits to the algorithm node,
* Directed edges from the algorithm node to the output bits.
}
\examples{
g <- convert_algo_base_to_igraph(node, ...);

}
