% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_any_to_logical_vector.R
\name{convert_any_to_logical_vector}
\alias{convert_any_to_logical_vector}
\title{convert_any_to_logical_vector}
\usage{
convert_any_to_logical_vector(input = NULL, dim = NULL, ...)
}
\arguments{
\item{input}{An object or scalar (any supported class)}

\item{dim}{(conditional) A fixed dimension for the resulting binary number (integer)}

\item{...}{For future use}
}
\value{
The resulting binary number represented as a logical vector (logical vector)
}
\description{
Receives an input of an arbitrary supported class and returns a binary number represented by a logical vector.
NULL is converted to a logical vector of length 1 with a FALSE value.
}
\examples{
# R function style:
convert_any_to_logical_vector("10001");
convert_any_to_logical_vector(c(TRUE,FALSE,FALSE,FALSE,TRUE));
convert_any_to_logical_vector(17);

}
