% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_int_to_bnum.R
\name{convert_int_to_bnum}
\alias{convert_int_to_bnum}
\title{convert_int_to_bnum}
\usage{
convert_int_to_bnum(i = NULL, dim = NULL, ...)
}
\arguments{
\item{i}{An integer (integer)}

\item{dim}{The dimension of the binary number (integer)}
}
\value{
The resulting binary number (R6 class bnum)
}
\description{
Convert a native R integer value into a logical vector representation.
}
\examples{
# R function style:
convert_int_to_bnum(57, 8);

}
