% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_int_to_logical_vector.R
\name{convert_int_to_logical_vector}
\alias{convert_int_to_logical_vector}
\title{convert_int_to_logical_vector}
\usage{
convert_int_to_logical_vector(i = NULL, dim = NULL, ...)
}
\arguments{
\item{i}{A natural number (integer)}

\item{dim}{(conditional) The size (length) of the logical vector (integer)}
}
\value{
The logical vector (logical vector)
}
\description{
Convert a native R integer value into a binary number materialized by an R logical vector. \cr
See \code{\link[modular_binary_number]{modular binary number}} for a definition of binary number. \cr
See \code{\link{convert_any_to_logical_vector}} for the type generic equivalent function. \cr
Open question: should this function provide support for vectors of integers? If yes, what should it return? A matrix? \cr
}
\examples{
# R function style:
convert_int_to_logical_vector(137, 9);

}
