% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_int_to_position.R
\name{convert_int_to_position}
\alias{convert_int_to_position}
\title{convert_int_to_position}
\usage{
convert_int_to_position(i)
}
\description{
In a sequence of incremented binary numbers with integer values starting from 0 to n,
the index position of the binary number with inteteger value 0 is 1.
}
