% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nandtree_to_truthtable.R
\name{convert_nandtree_to_truthtable}
\alias{convert_nandtree_to_truthtable}
\title{Convert a NAND tree to a truth table.}
\usage{
convert_nandtree_to_truthtable(nandtree)
}
\arguments{
\item{nandtree}{A NandTree (R6 Class NandTree)}
}
\value{
The resulting truth table (R6 Class Truthtable)
}
\description{
This function builds a truth table from a NAND tree
by executing its algorithm with all possible inputs.
Of course, this function is highly inefficient
and may only be applied to low input dimensional NAND trees.
}
\examples{
# R function style:
convert_nandtree_to_truthtable(nandtree);

# R6 method style:
nandtree$convert_to_truthtable();

}
