% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_tt_constant_to_composite.R
\name{convert_tt_constant_to_composite}
\alias{convert_tt_constant_to_composite}
\title{convert_tt_constant_to_composite}
\usage{
convert_tt_constant_to_composite(algo)
}
\arguments{
\item{algo}{A truth table algorithm (R6 Class algo_tt)}
}
\value{
The equivalent composite algorithm (R6 Class composite)
}
\description{
Convert a constant truth table algorithm,
i.e. a truth table algorithm of input dimension 1,
and of arbitrary output dimension > 0,
and returns a composite algorithm
that reproduce the equivalent logic
but with only atomic NANDs and or atomic constants.
}
\examples{
...

}
