% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fun_copy_logic_algo_composite_to_algo_composite.R
\name{copy_logic_algo_composite_to_algo_composite}
\alias{copy_logic_algo_composite_to_algo_composite}
\title{copy_logic_algo_composite_to_algo_composite, algo_composite$copy_logic_to, algo_composite$copy_logic_from}
\usage{
copy_logic_algo_composite_to_algo_composite(source, target, ...)
}
\arguments{
\item{source}{A composite algorithm (R6 Class algo_composite)}

\item{target}{A composite algorithm (R6 Class algo_composite)}

\item{...}{For future usage}
}
\value{
N/A
}
\description{
Copies the internal logic of a composite algorithm on a different composite algorithm.
}
\examples{
# R function style:
a1 <- algo_xnor$new();
a2 <- algo_composite$new(input_dimension = 2, output_dimension = 1);
copy_logic_algo_composite_to_algo_composite(a1, a2);
a2$plot();

# R6 method style with copy from:
a3 <- algo_composite$new(input_dimension = 2, output_dimension = 1);
a3$copy_logic_from(a1);
a3$plot();

# R6 method style with copy to:
a4 <- algo_composite$new(input_dimension = 2, output_dimension = 1);
a1$copy_logic_to(a4);
a4$plot();

}
