% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_modular_binary_domain.R
\name{def_modular_binary_domain}
\alias{def_modular_binary_domain}
\title{Modular binary domain}
\description{
In the context of this package, a modular binary domain of dimension \eqn{d} and noted \eqn{\mathbb{B}_{d}} is defined as the ordered set of all possible modular binary numbers of size \eqn{d}.
}
\section{Notation}{

A modular binary domain of dimension \eqn{d} is noted:\eqn{\mathbb{B}_{d}}.
}

\section{Samples}{

\deqn{\mathbb{B}_{1} = \left\{0,1\right\}}
\deqn{\mathbb{B}_{2} = \left\{00,10,01,11\right\}}
\deqn{\mathbb{B}_{3} = \left\{000,100,010,110,001,101,011,111\right\}}
}

\seealso{
Other binaries: \code{\link{def_modular_binary_number}}
}
\concept{binaries}
