% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_modular_binary_number.R
\name{def_modular_binary_number}
\alias{def_modular_binary_number}
\title{Modular binary number}
\description{
In the context of this package, a modular binary number of dimension \eqn{d} and noted \eqn{n_{d}} is defined as an ordered set of \eqn{d} bits.
}
\section{Notation}{

A modular binary number of dimension \eqn{d} is noted:\eqn{n_{d}}.
}

\section{Samples}{

\deqn{a_{4} \in \mathbb{B}_{4}}
\deqn{a_{4} = \left\{0,1,0,1\right\}}
\deqn{b_{6} \in \mathbb{B}_{6}}
\deqn{b_{6} = \left\{0,0,0,0,0,0\right\}}
\deqn{c_{3} \in \mathbb{B}_{3}}
\deqn{c_{3} = \left\{1,1,0\right\}}
}

\seealso{
Other binaries: \code{\link{def_modular_binary_domain}}
}
\concept{binaries}
