% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_splitting_truth_tables.R
\name{def_splitting_truth_tables}
\alias{def_splitting_truth_tables}
\title{Definition: Splitting truth tables

In the context of this package, the splitting operation applied on a truth table \eqn{a} means:
\enumerate{
\item ...
\item ...
\item ...
}}
\description{
Definition: Splitting truth tables

In the context of this package, the splitting operation applied on a truth table \eqn{a} means:
\enumerate{
\item ...
\item ...
\item ...
}
}
\section{Notation}{

The concatenation operator is noted:\eqn{split}.
}

\section{Samples}{

\deqn{ 01 & 1010 = 011010 }
\deqn{ 000 & 00 = 00000 }
}

\concept{transformations}
