% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_dimensional_equivalence.R
\name{dimensional_equivalence}
\alias{dimensional_equivalence}
\title{Dimensional equivalence}
\description{
A transformation \eqn{t} has the dimensional equivalence property iif: \cr
\deqn{\forall a_{1},a_{2} \in \mathbb{A}: a_{1} = t(a_{2}) \implies dim_i(a_{1}) = dim_i(a_{2}), dim_o(a_{1}) = dim_o(a_{2})}
}
\seealso{
\code{\link{split}}
}
