% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dissolve_tt.R
\name{dissolve_tt}
\alias{dissolve_tt}
\title{dissolve_tt}
\usage{
dissolve_tt(algo, ...)
}
\arguments{
\item{...}{For future usage}

\item{source}{A truth table algorithm (R6 Class algo_tt)}
}
\value{
A composite algorithm (R6 Class algo_composite)
}
\description{
This transformation takes a composite or truth table in input and returns a composite with identical logic.
}
\examples{
...

}
