% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_aggregate_MoBiNum.R
\name{do_aggregate_MoBiNum}
\alias{do_aggregate_MoBiNum}
\title{do_aggregate_MoBiNum}
\usage{
do_aggregate_MoBiNum(...)
}
\arguments{
\item{...}{Modular binary numbers (MoBiNum)}
}
\value{
The result of the aggregation
}
\description{
...
}
\examples{
Function style:
n1 <- MoBiNum$new("0101");
n2 <- MoBiNum("11000");
n3 <- do_aggregate_MoBiNum(n1, n2);
print(n3);

R6 method style:
n4 <- n1$aggregate_with(n2);
print(n4);

}
