% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_change_id_algo_composite.R
\name{do_change_id_algo_composite}
\alias{do_change_id_algo_composite}
\title{do_change_id_algo_composite, algo_composite$do_change_id}
\usage{
do_change_id_algo_composite(algo, id, ...)
}
\arguments{
\item{algo}{A composite algorithm (R6 Class algo_composite)}

\item{id}{A unique identifier (character)}

\item{...}{For future usage}
}
\value{
N/A
}
\description{
Modifies the unique ID of a composite algorithm node.
}
\examples{
# R function style:
a1 <- algo_xnor$new();
print(a1$get_node_id());
do_change_id_algo_composite(a1, "a_stupid_id");
print(a1$get_node_id());

# R6 method style:
a2 <- algo_xnor$new();
print(a2$get_node_id());
a2$do_change_id("another_stupid_id");
print(a2$get_node_id());

}
