% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_compute_vertex_bit_id.R
\name{do_compute_vertex_bit_id}
\alias{do_compute_vertex_bit_id}
\title{do_compute_vertex_bit_id}
\usage{
do_compute_vertex_bit_id(type, number, ...)
}
\arguments{
\item{type}{The bit type: "i" or "o" (character)}

\item{number}{The bit number (integer)}

\item{...}{For future usage}
}
\value{
The bit name
}
\description{
Computes the "bit_id" attribute of igraph graphs in CompositeNode.
The naming scheme goes like this:
For input bit vertices: "i1", "i2", "i3", ...
For algo vertices: "" (empty string)
For output bit vertices: "o1", "o2", "o3", ...
This function is mainly used for internal purposes but could be useful for advanced manipulations.
This function is trivial but may prove useful in the future to enforce data quality validation rules.
}
\examples{
print(do_compute_vertex_bit_id("i", 5));
print(do_compute_vertex_bit_id("o", 1:4));

}
