% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_concatenate_MoBiNum.R
\name{do_concatenate_MoBiNum}
\alias{do_concatenate_MoBiNum}
\title{Function: do_concatenate_MoBiNum}
\usage{
do_concatenate_MoBiNum(...)
}
\arguments{
\item{...}{Modular binary numbers (MoBiNum)}
}
\value{
The result of the concatenation
}
\description{
Concatenates modular binary numbers.
}
\section{Definitions}{

\itemize{
\item{Concatenation of modular binary numbers (\link{def_concatenation_modular_binary_numbers})}
}
}

\examples{
Function style:
n1 <- MoBiNum$new("0101");
n2 <- MoBiNum("11000");
n3 <- do_concatenate_MoBiNum(n1, n2);
print(n3);

R6 method style:
n4 <- n1$concatenate_with(n2);
print(n4);

Infix operator style:
n5 <- n1 \%&\% n2;
print(n5);

}
\seealso{
Other concatenations: \code{\link{def_concatenation_modular_binary_numbers}}
}
\concept{concatenations}
