% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_algo_base_to_algo_tt.R
\name{do_convert_algo_base_to_algo_}
\alias{do_convert_algo_base_to_algo_}
\title{do_convert_algo_base_to_algo_}
\usage{
do_convert_algo_base_to_algo_(algo)
}
\arguments{
\item{algo}{An arbitrary algorithm (R6 Class algo_base)}
}
\value{
The equivalent truth-table-based algorithm (R6 Class algo_)
}
\description{
This function builds a truth table from an arbitrary algorithm
by executing it with all possible inputs.
Of course, this function is highly inefficient
and may only be applied to low input dimensional algorithms.
}
\examples{
# R function style:
a1 <- AlgoNOT$new();
a2 <- do_convert_algo_base_to_algo_(a1);
print(a2);

# R6 method style:
a3 <- a1$convert_to_algo_();
print(a3);

}
