% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_algo_composite_to_igraph.R
\name{do_convert_algo_composite_to_igraph}
\alias{do_convert_algo_composite_to_igraph}
\title{Convert an object of type algo_composite to an igraph object.}
\usage{
do_convert_algo_composite_to_igraph(node, ...)
}
\arguments{
\item{node}{The node that we want to graph (algo_composite)}

\item{...}{For future usage.}
}
\value{
A directed network graph representation of the node (igraph).
}
\description{
The igraph network will be composed of:
* One node per input bit,
* One node per algorithm,
* One node per output bit,
* Directed edges from the input bits to the algorithm node,
* Directed edges from the algorithm node to the output bits,
* Directed edges between sub-algorithms.
}
\examples{
g <- convert_algo_composite_to_igraph(node, ...);

}
