% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_AlgoTT_to_character_dataframe.R
\name{do_convert_algo_tt_to_character_dataframe}
\alias{do_convert_algo_tt_to_character_dataframe}
\title{do_convert_algo_tt_to_character_dataframe}
\usage{
do_convert_algo_tt_to_character_dataframe(algo_tt, ...)
}
\arguments{
\item{...}{For future usage}

\item{algo_}{A truth table alogirthm (R6 Class algo_)}
}
\value{
A nicely formatted ROxygen2 Tabular (character)
}
\description{
Takes an object of type algo_ and returns a character dataframe representation.
This format is both machine and human readable.
Useful to populate truth tables in ROxygen2 tabular format with sinew for package documentation.
}
\examples{
# Function style:
a1 <- algo_tt$new(input_dimension = 2, output_dimension = 3);
do_convert_algo_tt_to_character_dataframe(a1);

# R6 method style:
a1$convert_to_character_dataframe();

}
