% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_AlgoNode_to_AlgoTT.R
\name{do_convert_AlgoNode_to_AlgoTT}
\alias{do_convert_AlgoNode_to_AlgoTT}
\title{do_convert_AlgoNode_to_AlgoTT}
\usage{
do_convert_AlgoNode_to_AlgoTT(algo)
}
\arguments{
\item{algo}{An arbitrary algorithm (R6 Class AlgoNode)}
}
\value{
The equivalent truth-table-based algorithm (R6 Class AlgoTT)
}
\description{
This function builds a truth table from an arbitrary algorithm
by executing it with all possible inputs.
Of course, this function is highly inefficient
and may only be applied to low input dimensional algorithms.
}
\examples{
# R function style:
a1 <- AlgoNOT$new();
a2 <- do_convert_AlgoNode_to_AlgoTT(a1);
print(a2);

# R6 method style:
a3 <- a1$convert_to_AlgoTT();
print(a3);

}
