% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_any_to_bnum.R
\name{do_convert_any_to_bnum}
\alias{do_convert_any_to_bnum}
\title{do_convert_any_to_bnum}
\usage{
do_convert_any_to_bnum(input)
}
\arguments{
\item{input}{A modular binary number (logical vector, character vector, bnum)}
}
\value{
A modular binary number (bnum)
}
\description{
Convert a modular binary number represented in an arbitrary type to a modular binary number typed as bnum.
}
\examples{
do_convert_any_to_bnum("1101");
do_convert_any_to_bnum(c(TRUE,TRUE,FALSE,TRUE));
do_convert_any_to_bnum(bnum$new("1101"));

}
