% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_any_to_MoBiDum_character.R
\name{do_convert_any_to_MoBiDum_character}
\alias{do_convert_any_to_MoBiDum_character}
\title{do_convert_any_to_MoBiDum_character}
\usage{
do_convert_any_to_MoBiDum_character(input)
}
\arguments{
\item{input}{A modular binary number (logical vector, character vector, BiNum)}
}
\value{
A modular binary number (character)
}
\description{
A modular binary number typed as a character is a string composed of "0"s and "1"s.
The leftmost character (position = 1) being the least significant bit.
}
\details{
Convert a modular binary number of an arbitrary type to a modular binary number typed as character.
}
\examples{
# R function style:
do_convert_any_to_MoBiDum_character(input);

# R6 method style:
modbinum$convert_to_character();

}
