% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_convert_bdom_to_character_vector.R
\name{do_convert_bdom_to_character_vector}
\alias{do_convert_bdom_to_character_vector}
\title{do_convert_bdom_to_character_vector}
\usage{
do_convert_bdom_to_character_vector(binary_domain, ...)
}
\arguments{
\item{binary_domain}{A binary domain (R6 Class bdom)}

\item{...}{For future usage}
}
\value{
A vector where each item correspond to a binary number of the binary domain (character vector)
}
\description{
Takes a binary domain a returns a vector of character representations.
}
\examples{
# R function style:
b1 <- bdom$new(dimension = 3);
do_convert_bdom_to_character_vector(b1);

# R6 method style:
b1$do_convert_to_character_vector();

}
