% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fun_do_copy_logic_AlgoComposite_to_AlgoComposite.R
\name{do_copy_logic_AlgoComposite_to_AlgoComposite}
\alias{do_copy_logic_AlgoComposite_to_AlgoComposite}
\title{do_copy_logic_AlgoComposite_to_AlgoComposite, AlgoComposite$do_copy_logic_to, AlgoComposite$do_copy_logic_from}
\usage{
do_copy_logic_AlgoComposite_to_AlgoComposite(source, target, ...)
}
\arguments{
\item{source}{A composite algorithm (R6 Class AlgoComposite)}

\item{target}{A composite algorithm (R6 Class AlgoComposite)}

\item{...}{For future usage}
}
\value{
N/A
}
\description{
Copies the internal logic of a composite algorithm on a different composite algorithm.
}
\examples{
# R function style:
a1 <- AlgoXNOR$new();
a2 <- AlgoComposite$new(input_dimension = 2, output_dimension = 1);
do_copy_logic_AlgoComposite_to_AlgoComposite(a1, a2);
a2$do_plot();

# R6 method style with copy from:
a3 <- AlgoComposite$new(input_dimension = 2, output_dimension = 1);
a3$do_copy_logic_from(a1);
a3$do_plot();

# R6 method style with copy to:
a4 <- AlgoComposite$new(input_dimension = 2, output_dimension = 1);
a1$do_copy_logic_to(a4);
a4$do_plot();

}
