% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_execute_algo_composite.R
\name{do_execute_algo_composite}
\alias{do_execute_algo_composite}
\title{Execute the algorithm of a algo_composite on a given input.}
\usage{
do_execute_algo_composite(algo, input)
}
\arguments{
\item{algo}{A composite algorithm (R6 Class algo_composite)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
A algo_composite, by definition in this context, is an algorithm
that takes x bits as input,
applies to them a directed graph set of abstract node functions,
and returns y bits as output.
This function applies / executes / runs
that algorithm on a given input
and returns the corresponding output.
}
\examples{
# R function style:
execute_algorithm_algo_composite(algo, input);

# R6 method style:
algo$execute_algorithm(input);

}
