% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_execute_AlgoComposite.R
\name{do_execute_AlgoComposite}
\alias{do_execute_AlgoComposite}
\title{Execute the algorithm of a AlgoComposite on a given input.}
\usage{
# R function style:
execute_algorithm_AlgoComposite(algo, input);

# R6 method style:
algo$execute_algorithm(input);
}
\arguments{
\item{algo}{A composite algorithm (R6 Class AlgoComposite)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
A AlgoComposite, by definition in this context, is an algorithm
that takes x bits as input,
applies to them a directed graph set of abstract node functions,
and returns y bits as output.
This function applies / executes / runs
that algorithm on a given input
and returns the corresponding output.
}
