% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_plot_algo_composite.R
\name{do_plot_algo_composite}
\alias{do_plot_algo_composite}
\title{Plot the igraph of a algo_composite}
\usage{
do_plot_algo_composite(algo, ...)
}
\arguments{
\item{algo}{A composite algorithm (algo_composite).}

\item{...}{For future usage.}
}
\value{
A plotted network graph.
}
\description{
Plot the igraph of a algo_composite with adequate style, etc.
}
\examples{
a1 <- algo_xnor$new();
do_plot_algo_composite(a1);

}
