% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_do_plot_AlgoComposite.R
\name{do_plot_AlgoComposite}
\alias{do_plot_AlgoComposite}
\title{Plot the igraph of a AlgoComposite}
\usage{
do_plot_AlgoComposite(algo, ...)
}
\arguments{
\item{algo}{A composite algorithm (AlgoComposite).}

\item{...}{For future usage.}
}
\value{
A plotted network graph.
}
\description{
Plot the igraph of a AlgoComposite with adequate style, etc.
}
\examples{
a1 <- AlgoXNOR$new();
do_plot_AlgoComposite(a1);

}
