% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plot_AlgoNode.R
\name{do_plot_AlgoNode}
\alias{do_plot_AlgoNode}
\title{Plot the igraph of a AlgoNode}
\usage{
do_plot_AlgoNode(node, ...);
}
\arguments{
\item{...}{For future usage.}
}
\value{
A plotted network graph.
}
\description{
Plot the igraph of a AlgoNode with adequate style, etc.
}
