% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_exec_algo_0.R
\name{exec_algo_0}
\alias{exec_algo_0}
\title{Execute the 0 constant algorithm on no or a null input.}
\usage{
exec_algo_0(algo, input = NA)
}
\arguments{
\item{algo}{A 0 constant algorithm (R6 Class algo_0)}

\item{input}{(conditional) The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input, default type: logical vector)
}
\description{
...
}
\examples{
# R function style:
exec_algo_0(algo, input);

# R6 method style:
a <- algo_0$new();
a$exec(input);

}
