% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_exec_algo_tt.R
\name{exec_algo_tt}
\alias{exec_algo_tt}
\title{Execute a truth table algorithm on a given input.}
\usage{
exec_algo_tt(algo, input, ...)
}
\arguments{
\item{algo}{A truth table algorithm (R6 Class algo_tt)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
...
}
\examples{
# R function style:
exec_algo_tt(algo, input);

# R6 method style:
algo$exec(input);

}
