% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_algorithm_compositedigraph.R
\name{execute_algorithm_compositedigraph}
\alias{execute_algorithm_compositedigraph}
\title{Execute the algorithm of a composite digraph on a given input.}
\usage{
# R function style:
execute_algorithm_compositedigraph(compositedigraph, input);

# R6 method style:
compositedigraph$execute_algorithm(input);
}
\arguments{
\item{compositedigraph}{A composite tree (R6 Class CompositeDiGraph)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
A composite digraph, by definition in this context, is an algorithm
that takes x bits as input,
applies to them a directed graph set of abstract node functions,
and returns y bits as output.
This function applies / executes / runs
that algorithm on a given input
and returns the corresponding output.
}
