% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_algorithm_nandtree.R
\name{execute_algorithm_nandtree}
\alias{execute_algorithm_nandtree}
\title{Execute the algorithm of a NAND tree on a given input.}
\usage{
# R function style:
execute_algorithm_nandtree(nandtree, input);

# R6 method style:
nandtree$execute_algorithm(input);
}
\arguments{
\item{nandtree}{A NandTree (R6 Class NandTree)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
A NAND tree, by definition, is an algorithm
that takes x bits as input,
applies to them a tree (or directed graph) of NAND functions,
and returns y bits as output.
This function applies / executes / runs
that algorithm on a given input
and returns the corresponding output.
}
