% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_from_json.R
\name{from_json}
\alias{from_json}
\title{from_json}
\usage{
from_json(json, instance = NULL, ...)
}
\arguments{
\item{json}{The exported haricot JSON (json / character).}

\item{instance}{(Conditional) The target object. If missing or null, a new object will be created (R6 class instance).}
}
\value{
The object instance (R6 class)
}
\description{
Convert a JSON export of a haricot class to an instance object of that class. \cr
}
\examples{
...

}
