% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_node_guid.R
\name{get_node_guid}
\alias{get_node_guid}
\title{get_node_guid}
\usage{
get_node_guid()
}
\value{
A globally unique node identifier (character)
}
\description{
Return a pseudo-random globally unique identifier for nodes. \cr
}
\examples{
get_node_guid();

}
