% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_opt.R
\name{GLOBAL_OPTIONS}
\alias{GLOBAL_OPTIONS}
\title{get_opt}
\usage{
GLOBAL_OPTIONS(..., .__defaults = FALSE, .__reset = FALSE)
}
\arguments{
\item{...}{Overwrite package global option values with custom values.}

\item{o}{The unique name of the option to be retrieve (character)}
}
\value{
The option setting value.
}
\description{
Set an option setting. \cr
By default it will be a globally defined setting. \cr
But package users may tweak their preferences by passing new settings in ...
}
