% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_increase_output_bits.R
\name{increase_output_bits}
\alias{increase_output_bits}
\title{increase_output_bits}
\usage{
increase_output_bits(algo, n = NULL, ...)
}
\arguments{
\item{algo}{A composite algorithm (R6 Class algo_composite)}

\item{n}{(conditional, default = 1) The number of output bits to be added (integer)}
}
\value{
The composite algorithm (for function chaining)
}
\description{
Increase the number of output bits in a composite algorithm.
}
\examples{
# R function style:
a1 <- algo_composite$new(3,5);
increase_output_bits(a1, 2);

# R6 method style:
a2 <- algo_composite$new(3,5);
a2$increase_output_bits(2);

}
