% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haricot-package.R
\docType{package}
\name{<package-name>}
\alias{<package-name>}
\alias{<package-name>-package}
\title{haricot}
\description{
Amateur research work on algorithms. This package helps: \itemize{
\item compose,
\item manipulate,
\item execute,
\item analyse,
\item optimize
\item and visualize }
... algorithmic graphs.
}
\details{
Its key objects are:
\itemize{
\item The NAND atomic function,
\item Truth tables,
\item Logical gates,
\item Composite algorithms built from the above and other composite algorithms.
}
}
