% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_log.R
\name{log}
\alias{log}
\title{log}
\usage{
log(error_level = NULL, ...)
}
\arguments{
\item{error_level}{LOG_TRACE, LOG_INFO, LOG_WARN, LOG_ERROR, LOG_FATAL}

\item{...}{TBD}
}
\value{
Nothing
}
\description{
...
}
\examples{
...

}
