% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_splitting_truth_tables.R
\name{man_splitting_truth_tables}
\alias{man_splitting_truth_tables}
\title{Theory: Splitting truth tables

In the context of this package...}
\description{
Theory: Splitting truth tables

In the context of this package...
}
\seealso{
Other transformations: \code{\link{man_switching_algorithms}}
}
\concept{transformations}
