% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_switching_algorithms.R
\name{man_switching_algorithms}
\alias{man_switching_algorithms}
\title{Theory: Switching Algorithms

In the context of this package, ...}
\description{
Theory: Switching Algorithms

In the context of this package, ...
}
\seealso{
Other transformations: \code{\link{man_splitting_truth_tables}}
}
\concept{transformations}
