% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NandTree_FirstIdea.R
\docType{data}
\name{NandTree_FirstIdea}
\alias{NandTree_FirstIdea}
\title{NandTree_FirstIdea}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NandTree_FirstIdea
}
\description{
NandTree_FirstIdea
}
\keyword{datasets}
