% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_algo_base.R
\name{plot_algo_base}
\alias{plot_algo_base}
\title{Plot the igraph of a algo_base}
\usage{
plot_algo_base(node, ...)
}
\arguments{
\item{...}{For future usage.}
}
\value{
A plotted network graph.
}
\description{
Plot the igraph of a algo_base with adequate style, etc.
}
\examples{
a1 <- algo_nand$new();
plot_algo_base(a1);

}
