% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_algo_tt_inside.R
\name{plot_algo_tt_inside}
\alias{plot_algo_tt_inside}
\title{plot_algo_tt_inside}
\usage{
plot_algo_tt_inside(algo, interactive = FALSE, ...)
}
\arguments{
\item{algo}{The truth table algorithm to be plotted.}

\item{interactive}{(mentionned for consistency purpose, not yet supported).}

\item{...}{For future usage.}
}
\value{
A plotted truth table graph.
}
\description{
Plot the internal logic of a truth table algorithm.
}
\examples{
a1 <- algo_nand$new();
plot_algo_base(a1);

}
