% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_bdom.R
\name{plot_bdom}
\alias{plot_bdom}
\title{plot_bdom}
\usage{
plot_bdom(d, ...)
}
\arguments{
\item{d}{The binary domain to be plotted (R6 class bdom).}

\item{...}{For future usage.}
}
\value{
A plotted binary number.
}
\description{
Generate a visualization of a binary domain,
following the general look and feel of the haricot package.
}
\examples{
# R function style:
d <- bdom$new(3);
plot_bdom(d);

# R6 class method style:
d$plot();

}
