% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_bnum.R
\name{plot_bnum}
\alias{plot_bnum}
\title{plot_bnum}
\usage{
plot_bnum(n, ...)
}
\arguments{
\item{n}{The binary number to be plotted.}

\item{...}{For future usage.}
}
\value{
A plotted binary number.
}
\description{
Generate a visualization of a binary number,
following the general look and feel of the haricot package.
}
\examples{
# R function style:
n <- bnum$new("1010111000");;
plot_bnum(n);

# R6 class method style:
n$plot();

}
