% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CompositeAlgoInnerNode.R
\name{plot_CompositeAlgoInnerNode}
\alias{plot_CompositeAlgoInnerNode}
\title{Plot the igraph of a CompositeAlgoInnerNode}
\usage{
plot_CompositeAlgoInnerNode(node, ...);
}
\arguments{
\item{...}{For future usage.}
}
\value{
A plotted network graph.
}
\description{
Plot the igraph of a CompositeAlgoInnerNode with adequate style, etc.
}
