% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_predecessors.R
\name{predecessors}
\alias{predecessors}
\title{predecessors}
\usage{
predecessors(composite, component, ...)
}
\arguments{
\item{composite}{The composite within which we want to perform this operation (R6 class algo_composite).}

\item{component}{The component algorithm whose predecessors we want to list (R6 class algo_base | character vector of the component ID).}
}
\value{
The list of predecessor algorithms (list of algo_base items).
}
\description{
In the context of a composite algorithm,
return the list of immediate predecessor algorithms
to a given component algorithm. \cr
\cr
If the component is a constant, the list of predecessors will be empty. \cr
\cr
Among the predecessors, we may find the composite if the component
is directly linked to one of the composite input bits. \cr
\cr
\strong{Notation} \cr
\deqn{N^{-}(x)} \cr
}
\examples{
...

}
