% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_remove_component.R
\name{remove_component}
\alias{remove_component}
\title{remove_component}
\usage{
remove_component(composite, component, ...)
}
\arguments{
\item{composite}{The composite algorithm (algo_composite)}

\item{...}{For future usage}

\item{old}{The component algorithm (algo_base)}
}
\value{
The composite algorithm with the component remove (algo_composite)
}
\description{
Remove a component algorithm from a composite algorithm.
}
\examples{
# ...

}
