% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample_algo_truthtable_json.R
\docType{data}
\name{sample_algo_tt_json}
\alias{sample_algo_tt_json}
\title{sample_algo_tt_json.RData}
\format{JSON materialized as character.}
\usage{
data(sample_algo_tt_json);
}
\description{
A sample JSON exportation of a random truth table algorithm.
}
\examples{
# Load the sample data
data(sample_algo_tt_json, package = "haricot");
print(haricot::sample_algo_tt_json);
a1 <- algo_tt$new();
a1$from_json(haricot::sample_algo_tt_json);
a1$exec("1010");

}
\keyword{datasets}
