% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_set_component.R
\name{set_component}
\alias{set_component}
\title{set_component}
\usage{
set_component(composite, node, ...)
}
\arguments{
\item{composite}{The composite algorithm we want to manipulate (algo_composite).}

\item{node}{The component algorithm we want to inject (algo_base).}
}
\value{
The modified composite algorithm (algo_composite).
}
\description{
Set a component algorithm in a composite algorithm.
}
\examples{
...

}
