% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_split_algo_tt.R
\name{split_algo_tt}
\alias{split_algo_tt}
\title{split_algo_tt}
\usage{
split_algo_tt(algo, ...)
}
\arguments{
\item{algo}{... (algo_tt)}

\item{...}{For future usage}
}
\value{
A composite algorithm that has an identitcal truth table to the input algorithm (algo_composite)
}
\description{
Decompose a truth table algorithm into two sub-truth table algorithm with a switch.
}
\examples{
...

}
