% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_swap.R
\name{swap}
\alias{swap}
\title{swap}
\usage{
swap(composite, old, new, ...)
}
\arguments{
\item{composite}{The composite algorithm where we want to operate the substitution (algo_composite)}

\item{old}{The old composant algorithm that we want to remove from the composite (algo_base)}

\item{new}{The new composant algorithm that we want to introduce in the composite (algo_base)}

\item{...}{For future usage}
}
\value{
The composite algorithm with the substitution.
}
\description{
...
}
\examples{
# ...

}
