% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_switch.R
\name{switch_algo}
\alias{switch_algo}
\title{switch_algo}
\usage{
switch_algo(algo_0, algo_1, ...)
}
\arguments{
\item{algo_0}{A component algorithm to be switched to when e == 0 (algo_base)}

\item{algo_1}{A component algorithm to be switched to when e == 1 (algo_base)}

\item{...}{For future usage}
}
\value{
A component algorithm whose inner logic switches algo_0 and algo_1 based on e
}
\description{
...
}
\examples{
# Switch two simple algos:
require(haricot);
algo_0 <- algo_01$new();
algo_0$plot();
algo_1 <- algo_10$new();
algo_1$plot();
algo_switch <- switch_algo(algo_0, algo_1);
algo_switch$exec("00");
algo_switch$exec("10");
algo_switch$exec("01");
algo_switch$exec("11");
algo_switch$plot();

}
