% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_to_json.R
\name{to_json}
\alias{to_json}
\title{to_json}
\usage{
to_json(o, ...)
}
\arguments{
\item{o}{An object from one of the haricot package R6 classes (R6 Class)}
}
\value{
JSON
}
\description{
Convert an object from a haricot class to JSON.
}
\examples{
...

}
