% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_transforming_truthtables_into_dag.R
\name{transform_truthtables_into_dag}
\alias{transform_truthtables_into_dag}
\title{Transformation: truth tables to DAG}
\description{
Transformation: truth tables to DAG
}
\examples{
# Create a random truth table algorithm with input dimension 8 and output dimension 8:
truthtable <- algo_tt$new(8,8)$do_randomize_outputs();
# Display the upper part of the truth table in the console:
cat(substr(truthtable$get_prettystring(),1,200));
# Transform the truth table to a DAG algorithm:
dag <- atomize(truthtable);
#

}
