% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_splitting_truth_tables.R
\name{unswitching_truth_tables}
\alias{unswitching_truth_tables}
\title{Theory: Unswitching truth tables

In the context of this package...}
\description{
Theory: Unswitching truth tables

In the context of this package...
}
\concept{transformations}
