% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_baptize_algo_bit.R
\name{baptize_algo_bit}
\alias{baptize_algo_bit}
\title{baptize_algo_bit}
\usage{
baptize_algo_bit(type, number = NULL, ...)
}
\arguments{
\item{type}{The bit type: "i" for input or "o" for output (character)}

\item{number}{The bit number (integer)}

\item{...}{For future usage}
}
\value{
The bit name
}
\description{
Assign a proper name for a bit in an algo.
The naming scheme goes like this:
\itemize{
\item{For input bits: \code{"i1", "i2", "i3", ...}}
\item{For output bits: \code{"o1", "o2", "o3", ...}}
\item{For non-bits (e.g. the "bit" attribute of an algo vertex in an igraph): \code{"x"} (empty string)}
}
Any programmatic manipulation of bits should avoid to implement bit names directly and rather rely on this function to guarantee upward compatibility with future evolutions of the naming scheme.
}
\examples{
print(baptize_algo_bit("i", 5));
print(baptize_algo_bit("o", 1:4));

}
