% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_convert_bdom_to_character_vector.R
\name{convert_bset_to_character_vector}
\alias{convert_bset_to_character_vector}
\title{convert_bset_to_character_vector}
\usage{
convert_bset_to_character_vector(binary_domain, ...)
}
\arguments{
\item{binary_domain}{A binary domain (R6 Class bset)}

\item{...}{For future usage}
}
\value{
A vector where each item correspond to a binary number of the binary domain (character vector)
}
\description{
Takes a binary domain a returns a vector of character representations.
}
\examples{
# R function style:
b1 <- bset$new(dimension = 3);
convert_bset_to_character_vector(b1);

# R6 method style:
b1$convert_to_character_vector();

}
