% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_algorithm_nandtree.R
\name{execute_algorithm_nandtree}
\alias{execute_algorithm_nandtree}
\title{Execute (run) the Algorithm of a NANDTree on a given input.}
\usage{
# R function
execute_algorithm_nandtree(nandtree, input);
\cr
# R6 method
nandtree$execute_algorithm(input);
}
\arguments{
\item{nandtree}{A NandTree (R6 Class TruthTable)}

\item{input}{The input bits (logical vector | character vector of "0"s and "1"s | R6 Class BinaryNumber)}
}
\value{
The corresponding output (same type than input)
}
\description{
A NandTree, by definition, is an algorithm
that takes X bits as input,
and returns Y bits as output.
This function applies / executes / runs
that algorithm on a given input
and returns the corresponding output.
}
